+++ lib/rubocop/cop/rspec/contain_exactly.rb
@@ -3,7 +3,11 @@
 module RuboCop
   module Cop
     module RSpec
-      # Prefer `match_array` when matching array values.
+      # Checks where `contain_exactly` is used.
+      #
+      # This cop checks for the following:
+      # - Prefer `match_array` when matching array values.
+      # - Prefer `be_empty` when using `contain_exactly` with no arguments.
       #
       # @example
       #   # bad
@@ -14,23 +18,48 @@ module RuboCop
       #
       #   # good
       #   it { is_expected.to contain_exactly(content, *array) }
+      #
+      #   # bad
+      #   it { is_expected.to contain_exactly }
+      #   it { is_expected.to contain_exactly() }
+      #
+      #   # good
+      #   it { is_expected.to be_empty }
+      #   it { is_expected.to be_empty }
+      #
       class ContainExactly < Base
         extend AutoCorrector
 
         MSG = 'Prefer `match_array` when matching array values.'
+        MSG_EMPTY_COLLECTION =
+          'Prefer `be_empty` when matching an empty collection.'
         RESTRICT_ON_SEND = %i[contain_exactly].freeze
 
         def on_send(node)
-          return unless node.each_child_node.all?(&:splat_type?)
-
-          add_offense(node) do |corrector|
-            autocorrect(node, corrector)
+          if node.arguments.empty?
+            check_empty_collection(node)
+          else
+            check_populated_collection(node)
           end
         end
 
         private
 
-        def autocorrect(node, corrector)
+        def check_empty_collection(node)
+          add_offense(node, message: MSG_EMPTY_COLLECTION) do |corrector|
+            corrector.replace(node.source_range, 'be_empty')
+          end
+        end
+
+        def check_populated_collection(node)
+          return unless node.each_child_node.all?(&:splat_type?)
+
+          add_offense(node) do |corrector|
+            autocorrect_for_populated_array(node, corrector)
+          end
+        end
+
+        def autocorrect_for_populated_array(node, corrector)
           arrays = node.arguments.map do |splat_node|
             splat_node.children.first
           end