+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master (Unreleased)
 
 - Fix a false positive for `RSpec/PendingWithoutReason` when pending/skip has a reason inside an example group. ([@ydah])
+- Add support for `RSpec/ContainExactly` when calling `match_array` with no arguments. ([@ydah])
 
 ## 2.19.0 (2023-03-06)